impression/pyproject.toml
[tool.poetry]
name = "impression"
version = "0.1.0"
description = ""
authors = ["Ricardo Matsui"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
inky = "^2.2.1"
pillow = "^11.3.0"
flask = "^3.1.2"
gunicorn = "^23.0.0"
rpi-gpio = "^0.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
